UNOMI-508 : refacto to allow to use simple scripts
